+++ src/KeyworkerProfile/components/KeyworkerProfile.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { properCaseName } from "../stringUtils";
+import { properCaseName } from "../../stringUtils";
 import PropTypes from 'prop-types';
 import { Link } from "react-router-dom";
 
@@ -56,8 +56,8 @@ class KeyworkerProfile extends Component {
     return (
       <div>
         <div className="pure-g">
-          <div className="pure-u-md-8-12">
-            <Link id={`search_again_link`} title="Search again link" className="link" to="/keyworkerSearch" >&lt; Search again</Link>
+          <div className="pure-u-md-8-12 padding-top">
+            <Link id={`search_again_link`} title="Search again link" className="link" to="/keyworker/search" >&lt; Search again</Link>
             <h1 className="heading-large">Profile for {keyworkerDisplayName}</h1>
           </div>
           {renderContent}
