+++ src/KeyworkerProfile/components/KeyworkerSearch.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import './index.scss';
+import '../index.scss';
 import PropTypes from 'prop-types';
+import { withRouter } from 'react-router';
 
 class KeyworkerSearch extends Component {
   render () {
@@ -10,7 +11,7 @@ class KeyworkerSearch extends Component {
           <div className="padding-top padding-left padding-right padding-bottom-large">
             <label className="form-label" htmlFor="seachText">key worker name</label>
             <input type="text" className="form-control" id="search-text" name="searchText" value={this.props.searchText} onChange={this.props.handleSearchTextChange}/>
-            <button className="button margin-left" onClick={() => this.props.gotoNext()}>Search ></button>
+            <button className="button margin-left" onClick={() => { this.props.handleSearch(this.props.history);}}>Search ></button>
           </div>
         </div>
       </div>
@@ -20,8 +21,12 @@ class KeyworkerSearch extends Component {
 
 KeyworkerSearch.propTypes = {
   searchText: PropTypes.string,
-  gotoNext: PropTypes.func.isRequired,
+  history: PropTypes.object,
+  handleSearch: PropTypes.func.isRequired,
   handleSearchTextChange: PropTypes.func.isRequired
 };
 
-export default KeyworkerSearch;
+const KeyworkerSearchWithRouter = withRouter(KeyworkerSearch);
+
+export { KeyworkerSearch };
+export default KeyworkerSearchWithRouter;
