+++ src/KeyworkerProfile/tests/keyworkerProfile.test.js
@@ -1,16 +1,15 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import KeyworkerSearchPage from "./index";
+import KeyworkerSearchPage from "../components/KeyworkerSearchPage";
 
 Enzyme.configure({ adapter: new Adapter() });
 
 
 describe('Keyworker search page component', () => {
   it('should render component correctly', async () => {
-    const component = shallow(<KeyworkerSearchPage handleSearchTextChange={jest.fn()} gotoNext={jest.fn()}/>);
-    expect(component.find('KeyworkerSearch').exists()).toEqual(true);
-    expect(component.find('div').at(1).text()).toContain('Search for a key worker');
+    const component = shallow(<KeyworkerSearchPage handleSearchTextChange={jest.fn()} handleSearch={jest.fn()}/>);
+    expect(component.text()).toContain('Search for a key worker');
   });
 });
